backend/dev_requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.5
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
future==0.18.2
idna==2.8
IPy==1.0
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.10
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
